+++ src/prefect/CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to
 
 - Implement t1 workflow
 - Implement i1 workflow
+- Integrate Great-Expectations
 
 ### Changed